+++ example/src/components/RowControls.jsx
@@ -5,12 +5,14 @@ import React from 'react';
 import {
 	Text,
 	View,
+	PixelRatio,
 } from 'react-native';
 
 /**
  * Internal dependencies
  */
 import Button from './Button.jsx';
+import ModalControl from './ModalControl.jsx';
 
 export const rowBtnStyle = {
     marginRight: 10,
@@ -23,6 +25,7 @@ export const ControlWrapper = ( {
     style,
     label,
     value,
+    valueMinWidth,
     textAppend,
     children,
 } ) => {
@@ -33,7 +36,7 @@ export const ControlWrapper = ( {
         ...containerStyle,
     } }>
         <Text style={ {...style, marginRight: 10, minWidth: 100 } }>{ label }:</Text>
-        { value !== null && <Text style={ {...style, marginRight: 10, minWidth: 30  } }>{ value }</Text> }
+        { value !== null && <Text style={ {...style, marginRight: 10, minWidth: valueMinWidth || 30  } }>{ value }</Text> }
         { children }
         { textAppend && <Text style={ { ...style, marginLeft: 10 }} >{ textAppend }</Text> }
     </View>;
@@ -48,12 +51,14 @@ export const ButtonControl = ( {
     buttonLabel,
     buttonStyle,
     onPress,
+    valueMinWidth,
 } ) => {
     return <ControlWrapper
         containerStyle={ containerStyle }
         style={ style }
         label={ label }
         textAppend={ textAppend }
+        valueMinWidth={ valueMinWidth }
     >
         <Button
             style={ { ...rowBtnStyle, ...buttonStyle } }
@@ -70,6 +75,7 @@ export const PlusMinusControl = ( {
     promiseQueueState,
     label,
     value,
+    valueMinWidth,
     minValue,
     setValue,
     step,
@@ -81,6 +87,7 @@ export const PlusMinusControl = ( {
         style={ style }
         label={ label }
         value={ value }
+        valueMinWidth={ valueMinWidth }
         textAppend={ textAppend }
     >
         <Button
@@ -98,3 +105,65 @@ export const PlusMinusControl = ( {
 
     </ControlWrapper>;
 };
+
+export const EventRowControl = ( {
+    style,
+    promiseQueueState,
+    mapViewNativeNodeHandle,
+    layerUuid,
+    width,
+    valueMinWidth,
+    mapHeight,
+    module,
+} ) => <ControlWrapper
+    containerStyle={ { marginBottom: 10 } }
+    style={ style }
+    label={ 'Events' }
+    value={ '' }
+    valueMinWidth={ valueMinWidth }
+>
+    <Button
+        style={ {
+            marginRight: 10,
+            width: 81,
+            textAlign: 'center',
+        } }
+        disabled={ promiseQueueState > 0 || ! layerUuid }
+        onPress={ () => {
+            if ( mapViewNativeNodeHandle && layerUuid ) {
+                module.triggerEvent(
+                    mapViewNativeNodeHandle,
+                    layerUuid,
+                    PixelRatio.getPixelSizeForLayoutSize( width ) / 2,
+                    PixelRatio.getPixelSizeForLayoutSize( mapHeight ) / 2
+                ).catch( err => console.log( 'ERROR', err ) );
+            }
+        } }
+        title={ 'trigger' }
+    />
+
+    <ModalControl
+        style={ style }
+        buttonStyle={ rowBtnStyle }
+        buttonLabel={ '?' }
+        headerLabel={ 'Events' }
+        disabled={ promiseQueueState > 0 }
+    >
+        <Text style={ {...style, marginBottom: 10} }>
+            This layer supports press, longPress and doubleTab events.
+        </Text>
+        <Text style={ {...style, marginBottom: 10} }>
+            Furthermore an event can be triggered at any map position, here the center.
+        </Text>
+        <Text style={ {...style, marginBottom: 10} }>
+            The gesture buffer distance can be controlled.
+        </Text>
+        <Text style={ {...style, marginBottom: 10} }>
+            Gesture events slow down the UI. If the layer doesn't have onPress|onLongPress|onDoubleTap functions, gesture support is disabled. Events can still be triggered.
+        </Text>
+        <Text style={ style }>
+            The response includes: event distance to the path, nearest point at path and the event position.
+        </Text>
+    </ModalControl>
+
+</ControlWrapper>;