+++ src/components/LayerPathSlopeGradient.tsx
@@ -9,13 +9,15 @@ import { useEffect, useState } from 'react';
 import useRefState from '../compose/useRefState';
 import promiseQueue from '../promiseQueue';
 import { MapLayerPathSlopeGradientModule } from '../nativeMapModules';
-import type { ResponseInclude, Location, LocationExtended, Bounds, ResponseBase } from '../types';
+import type { ResponseInclude, Location, LocationExtended, GeometryStyle, Bounds, ResponseBase } from '../types';
+import type { LayerPathGestureResponse } from './LayerPath';
+import { NativeEventEmitter } from 'react-native';
 
 const Module = MapLayerPathSlopeGradientModule;
 
 export type GradientColors = [number, `#${string}`][];
 
-export interface LayerPathSlopeGradientResponse extends ResponseBase  {
+export interface LayerPathSlopeGradientResponse extends ResponseBase {
 	coordinates?: LocationExtended[];
 	coordinatesSimplified?: LocationExtended[];
 	bounds?: Bounds;
@@ -27,14 +29,20 @@ export type LayerPathSlopeGradientProps = {
 	filePath?: null | `/${string}` | `content://${string}`;
 	positions?: Location[];
 	responseInclude?: ResponseInclude;
+	gestureScreenDistance?: number;
+	simplificationTolerance?: number;
 	slopeColors?: GradientColors;
-	strokeWidth?: number;
+	style?: GeometryStyle;
 	slopeSimplificationTolerance?: number;
 	flattenWindowSize?: number;	// should be odd and greater 5.
 	onRemove?: null | ( ( response: ResponseBase ) => void );
 	onCreate?: null | ( ( response: LayerPathSlopeGradientResponse ) => void );
 	onChange?: null | ( ( response: LayerPathSlopeGradientResponse ) => void );
 	onError?: null | ( ( err: any ) => void );
+	onPress?: null | ( ( response: LayerPathGestureResponse ) => void );
+	onLongPress?: null | ( ( response: LayerPathGestureResponse ) => void );
+	onDoubleTap?: null | ( ( response: LayerPathGestureResponse ) => void );
+	onTrigger?: null | ( ( response: LayerPathGestureResponse ) => void );
 };
 
 // 0	never include in response.
@@ -66,20 +74,30 @@ const sortSlopeColors = ( slopeColors : GradientColors ) : GradientColors => [..
 	return 0;
 } );
 
+const defaultStyle : GeometryStyle = {
+	strokeWidth: 4,
+}
+
 const LayerPathSlopeGradient = ( {
 	nativeNodeHandle,
 	positions = [],
 	filePath,
-	strokeWidth = 4,
+	style = defaultStyle,
+	simplificationTolerance = 0,
 	slopeColors = slopeColorsDefault,
 	slopeSimplificationTolerance = 7,
 	flattenWindowSize = 9,
 	responseInclude = responseIncludeDefaults,
+	gestureScreenDistance = 20,
 	onCreate,
 	onRemove,
 	onChange,
 	reactTreeIndex,
 	onError,
+	onPress,
+	onLongPress,
+	onDoubleTap,
+	onTrigger,
 } : LayerPathSlopeGradientProps ) => {
 
 	// @ts-ignore
@@ -88,11 +106,13 @@ const LayerPathSlopeGradient = ( {
 	const [triggerCreateNew, setTriggerCreateNew] = useState<null | number>( null );
 
 	positions = positions || [];
-	strokeWidth = strokeWidth > 0 ? Math.round( strokeWidth ) : 4;
 	slopeSimplificationTolerance = slopeSimplificationTolerance >= 0 ? Math.round( slopeSimplificationTolerance ) : 7;
 	flattenWindowSize = flattenWindowSize % 2 != 0 && flattenWindowSize > 5 ? flattenWindowSize : 9;
 	slopeColors = sortSlopeColors( slopeColors.length > 0 ? slopeColors : slopeColorsDefault );
 	responseInclude = { ...responseIncludeDefaults, ...responseInclude };
+	style = {...defaultStyle, ...style };
+
+	const supportsGestures = !! onPress || !! onLongPress || !! onDoubleTap;
 
 	const createLayer = () => {
 		setUuid( false );
@@ -101,11 +121,14 @@ const LayerPathSlopeGradient = ( {
 				nativeNodeHandle,
 				positions,
 				filePath,
-				strokeWidth,
+				style,
 				slopeColors,
 				slopeSimplificationTolerance,
 				flattenWindowSize,
 				responseInclude,
+				!! supportsGestures,
+				gestureScreenDistance,
+				simplificationTolerance,
 				reactTreeIndex
 			).then( ( response: LayerPathSlopeGradientResponse ) => {
 				setUuid( response.uuid );
@@ -143,17 +166,32 @@ const LayerPathSlopeGradient = ( {
 	useEffect( () => {
 		if ( nativeNodeHandle && uuid ) {
             promiseQueue.enqueue( () => {
-                return Module.updateStrokeWidth(
+                return Module.updateStyle(
                     nativeNodeHandle,
                     uuid,
-					strokeWidth,
+					style,
 					responseInclude,
 				).then( ( response: LayerPathSlopeGradientResponse ) => {
 					onChange ? onChange( response ) : null;
                 } ).catch( ( err: any ) => { console.log( 'ERROR', err ); onError ? onError( err ) : null } );
             } );
 		}
-	}, [strokeWidth] );
+	}, [Object.values( style ).join( '' )] );
+
+	useEffect( () => {
+		if ( nativeNodeHandle && uuid ) {
+            promiseQueue.enqueue( () => {
+                return Module.updateGestureScreenDistance(
+                    nativeNodeHandle,
+                    uuid,
+					gestureScreenDistance,
+					responseInclude,
+				).then( ( response: LayerPathSlopeGradientResponse ) => {
+					onChange ? onChange( response ) : null;
+                } ).catch( ( err: any ) => { console.log( 'ERROR', err ); onError ? onError( err ) : null } );
+            } );
+		}
+	}, [gestureScreenDistance] );
 
 	useEffect( () => {
 		if ( nativeNodeHandle && uuid ) {
@@ -161,7 +199,7 @@ const LayerPathSlopeGradient = ( {
                return Module.updateSlopeColors(
 					nativeNodeHandle,
 					uuid,
-					strokeWidth,
+					style,
 					slopeColors,
 					responseInclude,
 				).then( ( response: LayerPathSlopeGradientResponse ) => {
@@ -177,7 +215,7 @@ const LayerPathSlopeGradient = ( {
                 return Module.updateCoordinatesSimplified(
 					nativeNodeHandle,
 					uuid,
-					strokeWidth,
+					style,
 					slopeSimplificationTolerance,
 					flattenWindowSize,
 					responseInclude,
@@ -212,8 +250,41 @@ const LayerPathSlopeGradient = ( {
 			? [...positions].map( pos => pos.lng + pos.lat ).join( '' )
 			: null
 		),
+		simplificationTolerance,
 		filePath,
 		Object.keys( responseInclude ).map( key => key + responseInclude[key] ).join( '' ),
+	] );
+
+	useEffect( () => {
+		const eventEmitter = new NativeEventEmitter();
+		let eventListener = eventEmitter.addListener( 'PathSlopeGradientGesture', ( response : LayerPathGestureResponse ) => {
+			if ( response.uuid === uuid ) {
+				switch( response.type ) {
+					case 'doubleTap':
+						onDoubleTap ? onDoubleTap( response ) : null;
+						break;
+					case 'LongPress':
+						onLongPress ? onLongPress( response ) : null;
+						break;
+					case 'press':
+						onPress ? onPress( response ) : null;
+						break;
+					case 'trigger':
+						onTrigger ? onTrigger( response ) : null;
+						break;
+				}
+			}
+		} );
+		return () => {
+			eventListener.remove();
+		};
+	}, [
+		uuid,
+		!! supportsGestures,
+		onDoubleTap,
+		onLongPress,
+		onPress,
+		onTrigger,
 	] );
 
 	return null;